87495-using sqlview to get datalocks

Related work items: #87495

+++ src/ESFA.DC.DataMatch.ReportService.Model/DASPayments/DataLockValidationError.cs
@@ -18,7 +18,7 @@ namespace ESFA.DC.DataMatch.ReportService.Model.DASPayments
 
         public string Collection { get; set; }
 
-        public byte CollectionPeriod { get; set; }
+        public int CollectionPeriod { get; set; }
 
         public DateTime LastSubmission { get; set; }
     }
